new function: predict scene
